pyproject.toml
[tool.poetry]
name = "sudopython"
version = "0.1.0"
description = ""
authors = ["Amirouche <amirouche@hyper.dev>"]
license = "LGPLv2.1"

[tool.poetry.dependencies]
python = "^3.8"
gunicorn = "^20.0.4"
lsm-db = "^0.6.4"
foundationdb = "^6.2.10"
Unidecode = "^1.1.2"
html2text = "^2020.1.16"
ulid-py = "^1.1.0"
PyStemmer = "^2.0.1"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]
jedi = "0.17.1"
debug = "^0.3.2"
ipython = "^7.19.0"
Pympler = "^0.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
